+++ src/core/leoFind.ts
@@ -1501,7 +1501,7 @@ export class LeoFind {
 
                     g.app.gui.loadSearchSettings();
 
-                    return;
+                    return w_result;
 
                 }
             });
@@ -1679,8 +1679,7 @@ export class LeoFind {
                 `changed ${count} instances${count} ` + `in ${t2 - t1} sec.`
             );
         }
-        // c.recolor(); // ? NEEDED ?
-        // c.redraw(p); // ? NEEDED ?
+
         c.redraw(new_p);
         this.restore(saveData);
 
@@ -2495,7 +2494,7 @@ export class LeoFind {
      */
     private put_link(line: string, line_number: number, v: VNode): void {
         const c = this.c;
-        // const log = c.frame.log // UNAVAILABLE IN LEOJS
+        // const log = c.frame.log // UNAVAILABLE IN Leo-Web
         this.total_links += 1;
         if (this.total_links > 100) {
             return;
@@ -2776,7 +2775,7 @@ export class LeoFind {
                 }
             }
             return true;
-        }
+        };
 
         // Replace the placeholder text.
         settings.find_text = settings.find_text.replace('<find pattern here>', '');